Moved ModelViews to separate folder and namespace
