+++ FlatFleet/RecoverPasswordPage.xaml.cs
@@ -1,3 +1,5 @@
+using FlatFleet.ViewModels;
+
 namespace FlatFleet;
 
 public partial class RecoverPasswordPage : ContentPage
