+++ FlatFleet/ViewModels/TermsOfServiceViewModel.cs
@@ -1,12 +1,12 @@
 ﻿using Navigation;
 using System.Windows.Input;
 
-namespace FlatFleet
+namespace FlatFleet.ViewModels
 {
     public class TermsOfServiceViewModel : BindableObject
     {
         public ICommand OnSignInCommand { get; }
-        public TermsOfServiceViewModel() 
+        public TermsOfServiceViewModel()
         {
             OnSignInCommand = new Command(OnSignIn);
         }
